requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
markupsafe==2.1.3
psycopg2-binary
python-dotenv
Authlib
requests
